distage-extension-plugins: Fix StaticPluginLoader on Scala.js
- It didn't work because it was using a stub plugin loader, instead of classgraph plugin loader
