+++ distage/distage-extension-config/.js/src/main/scala/izumi/distage/config/codec/DIConfigReader.scala
@@ -7,26 +7,29 @@ import scala.util.Try
 import izumi.distage.config.DistageConfigImpl
 
 trait DIConfigReader[A] extends AbstractDIConfigReader[A] {
-  final def decodeConfig(config: DistageConfigImpl): Try[A] = ???
+  def decodeConfig(config: DistageConfigImpl): Try[A] = ???
 
-  final def decodeConfig(path: String)(config: DistageConfigImpl)(implicit tag: Tag[A]): A = {
+  def decodeConfig(path: String)(config: DistageConfigImpl)(implicit tag: Tag[A]): A = {
     ???
   }
 
-  final def map[B](f: A => B): DIConfigReader[B] = ???
+  def map[B](f: A => B): DIConfigReader[B] = ???
 
-  final def flatMap[B](f: A => DIConfigReader[B]): DIConfigReader[B] = ???
+  def flatMap[B](f: A => DIConfigReader[B]): DIConfigReader[B] = ???
 
-  final def decodeConfigWithDefault(path: String)(default: => A)(config: DistageConfigImpl)(implicit tag: Tag[A]): A = {
+  def decodeConfigWithDefault(path: String)(default: => A)(config: DistageConfigImpl)(implicit tag: Tag[A]): A = {
     ???
   }
 }
 
 object DIConfigReader {
-  @inline def apply[T: DIConfigReader]: DIConfigReader[T] = ???
+  @inline def apply[T: DIConfigReader]: DIConfigReader[T] = implicitly
 
   def derived[T: ClassTag]: DIConfigReader[T] = ???
 
-  implicit final def todo[T: ClassTag]: DIConfigReader[T] = ???
-
+  implicit final def todo[T: ClassTag]: DIConfigReader[T] = new DIConfigReader[T] {
+    override def decodeConfig(config: DistageConfigImpl): Try[T] = ???
+    override def decodeConfig(path: String)(config: DistageConfigImpl)(implicit tag: Tag[T]): T = ???
+    override def decodeConfigWithDefault(path: String)(default: => T)(config: DistageConfigImpl)(implicit tag: Tag[T]): T = ???
+  }
 }
