+++ distage/distage-extension-plugins/src/main/scala-3/izumi/distage/plugins/StaticPluginLoader.scala
@@ -1,9 +1,8 @@
 package izumi.distage.plugins
 
 import izumi.distage.model.definition.ModuleBase
-import izumi.distage.plugins.load.{LoadedPlugins, PluginLoaderDefaultImpl}
+import izumi.distage.plugins.load.{LoadedPlugins, PluginLoaderClassgraphImpl}
 
-import scala.compiletime.error
 import scala.quoted.{Expr, Quotes, Type}
 
 /** Scan the specified package *at compile-time* for classes and objects that inherit [[PluginBase]]
@@ -24,7 +23,7 @@ object StaticPluginLoader {
     val loadedPlugins = if (pluginPath == "") {
       LoadedPlugins.empty
     } else {
-      new PluginLoaderDefaultImpl().load(PluginConfig.packages(Seq(pluginPath)))
+      new PluginLoaderClassgraphImpl().load(PluginConfig.packages(Seq(pluginPath)))
     }
 
     val quoted = instantiatePluginsInCode[PluginBase](loadedPlugins.loaded)
