+++ distage/distage-extension-plugins/src/test/scala/com/github/pshirshov/test/plugins/ObjectTestPlugin.scala
