+++ distage/distage-extension-plugins/src/test/scala/com/github/pshirshov/test/sneaky/SneakyPlugin.scala
