+++ distage/distage-framework/src/test/scala/com/github/pshirshov/test/plugins/StaticTestRole.scala
